+++ Scripts/Data/QuestStep.cs
@@ -22,6 +22,12 @@ namespace RAXY.Quest
         [HideReferenceObjectPicker]
         public List<QuestAction> actions_OnEnter = new();
 
+        [TitleGroup("Step Actions")]
+        [SerializeReference]
+        [ListDrawerSettings(ShowIndexLabels = true, ListElementLabelName = "Label", Expanded = true)]
+        [HideReferenceObjectPicker]
+        public List<QuestAction> actions_OnComplete = new();
+
         [TitleGroup("Objectives")]
         [ListDrawerSettings(ShowIndexLabels = true, ListElementLabelName = "ObjectiveDescription", Expanded = true)]
         [HideReferenceObjectPicker]